+++ tests/test_kernels/fixed_kernel_address/src/bin/should_panic.rs
@@ -1,0 +1,20 @@
+#![no_std] // don't link the Rust standard library
+#![no_main] // disable all Rust-level entry points
+
+use bootloader_api::{entry_point, BootInfo};
+use test_kernel_fixed_kernel_address::BOOTLOADER_CONFIG;
+
+entry_point!(kernel_main, config = &BOOTLOADER_CONFIG);
+
+fn kernel_main(_boot_info: &'static mut BootInfo) -> ! {
+    panic!();
+}
+
+/// This function is called on panic.
+#[cfg(not(test))]
+#[panic_handler]
+fn panic(_info: &core::panic::PanicInfo) -> ! {
+    use test_kernel_fixed_kernel_address::{exit_qemu, QemuExitCode};
+
+    exit_qemu(QemuExitCode::Success);
+}